project1/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
idna==2.10
importlib-metadata==1.7.0
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
urllib3==1.25.10
zipp==3.1.0